requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.24
django-bootstrap3==9.1.0
django-bower==5.2.0
django-cors-headers==2.1.0
django-countries==5.0
django-datetime-widget==0.9.3
django-filter==2.0.0
django-rest-swagger==2.1.2
django-split-settings==0.1.3
djangorestframework==3.11.2
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.3
Markdown==2.6.2
MarkupSafe==1.1.1
nodeenv==0.13.2
openapi-codec==1.3.2
psycopg2==2.8.6
pytz==2019.2
requests==2.26.0
rest-framework-ember==1.3.2
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.26.6
